Implemented RandomCastlingRuleTrait
